+++ uninstall.php
@@ -38,6 +38,7 @@ $options = array(
 	'ati_ga4_missing_cid_policy',
 	'ati_ga4_dedup_ttl',
 	'ati_ga4_enable_submit_attempt',
+	'ati_ga4_lead_trigger',
 	'ati_ga4_debug_mode',
 	'ati_ga4_form_map',
 	'ati_ga4_token_salt',